feat: implement error handling with catch-table and protected-frame policies
