+++ pkg/vm/vm_errors_test.go
